fix packaging and relative imports
