+++ backend/__init__.py
@@ -1,0 +1,1 @@
+version="0.0.1"
