+++ backend/chats/__init__.py
@@ -1,0 +1,6 @@
+__all__ = [
+        'chat',
+        'models',
+        'services',
+        ]
+
