+++ backend/chats/models.py
@@ -41,7 +41,7 @@ class SuccessDTO(PydanticBaseModel):
     msg: Optional[str]
 
 
-Tortoise.init_models(["models"], "models")
+Tortoise.init_models(["chats.models"], "models")
 Chat = pydantic_model_creator(Chats, name="Chat")
 Message = pydantic_model_creator(Messages, name="Message")
 MessageIn = pydantic_model_creator(Messages, name="Message", exclude=["chat"])
